Some SonarQube cleanup

+++ aoc2019/aoc2019-kotlin/src/test/kotlin/de/havox_design/aoc2019/day13/Day13Test.kt
@@ -15,7 +15,7 @@ class Day13Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         CarePackagePart2(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
